libsplash Makefile: print fortran compiler
And version. This is useful, as there have been issues with gfortran
versions. gfortran-7 seems to have memory leak issues, whereas
gfortran-8 doesn't.

The memory leak was due to using the same name in libsplash.f90 for each
of the subroutines, even though we were renaming them on use from the
parent module.


Former-commit-id: fe0944b9d3d4892169d6b08164b89ae3956a199d