Update contributions and citation
Create CONTRIBUTING.md to explain how to contribute to Plonk. Added
CITATION.bib for Astrophysics Source Code Library citation.

Remove the contributing section of the documentation to not repeat the
same content in two places. Plus it is not really documenting the use of
the code which is what I expect user documentation to do.


Former-commit-id: 0746e71479e29c0f3093fe2f93cc105c995a3a48