WIP: Clean up interpolation function interfaces
Now we pass in the x, y, z coordinates as separate arrays. This is to
make it easier to use different coordinate systems. For example, x and y
could be R and Z in cylindrical coordinates.

We also pass in the x and y components of the vector quantity as
separate arrays. Again this is to make it easier to use different
coordinte systems.

Note this breaks the Visualization class, so this commit is a WIP.