Simplify Visualization class
This is a complete rewrite of the Visualization class. It simplifies the
code at the expense of making the user interface more complicated.
However, it opens up more options to the user.

Now, the user must specify the x, y, z coordinates as arrays, and the
scalar data and vector data are also passed in as arrays. The user
cannot specify the data to plot via a string. Perhaps such a function or
class could use Visualization for that style of plotting.

Given that the user inputs arrays, the processing of the data occurs
before reaching the Visualization object. Processing such as unit
conversion, selecting a subset of particles to plot, etc.

The VisualizationIterator is removed.