Regrid interpolated data using RectBivariateSpline
This is a bug fix. The grid that KDEpy.FFTKDE returns the data on is not
the same as that requested. We need to re-interpolate onto the expected
grid. This is fast, as the pixel grid is small, compared to the number
of particles, and it is regular, i.e. gridded, data.