WIP: Separate out reading files from Dump class
This is a big change, and a WIP. The structure of the sub-packages has
changed. The main change is that the data reader, for example, for
Phantom HDF5 files is separated from the Dump class which is an
abstract class for an SPH simulation snapshot, i.e. dump.

There are several changes remaining to implement:
- sink particles are not read
- all other modules that read dump objects need to be changed