Calculate dust quantities differently
The calculation depends on the dust method. Previously, for the 'dust as
separate sets of particles' method the dust quantities were incorrect.
In addition, the Phantom 'dustfrac' array is not really the dust
fraction. It's better not to read it as is but set it to be one if the
particle is a dust particle of the correct species or zero otherwise.