Refactor load_snap
Now load_snap is a method on Snap. This change abstracts some of the
loading of data away from the phantom.py reader and puts it in Snap
itself.

To write a new reader, three functions are required:

- snap_properties_and_units,
- snap_array_registry,
- snap_sink_registry.

These functions take an h5py.File object as an argument (and optionally
a name map dictionary to rename arrays on file) and a string data_source
to say where the data is from, e.g. 'phantom'.

The first function gets some properties and code units, e.g. from the
Phantom header. The second two functions each create a dictionary with
keys as names of arrays and values as a function that when called with a
Snap object returns that array.